Removed Heap, reworked Huffman for heapq
